+++ src/Utils/Point.hpp
@@ -6,9 +6,9 @@ namespace gsim {
 	/// @brief A struct representing a 2D vector.
 	struct Vector2 {
 		/// @brief The X dimension of the vector.
-		double x;
+		float x;
 		/// @brief The Y dimension of the vector.
-		double y;
+		float y;
 	};
 
 	/// @brief A struct representing a simulated point.
@@ -18,6 +18,6 @@ namespace gsim {
 		/// @brief The velocity of the point.
 		Vector2 vel;
 		/// @brief The mass of the point.
-		double mass;
+		float mass;
 	};
 }
